Reorganise react components
