+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/QueryEditor/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './QueryEditor';
